Refactor Ludo architecture: Centralize logic in GameState controller, create HomeSpotManager layout registry, and clean up UI components
